Revert "Add CMake support"
