+++ .gitignore
@@ -1,4 +1,2 @@
 build/
 a.out
-CMakeCache.txt
-CMakeFiles
