--- PreventInSouceBuilds.cmake.in
@@ -1,13 +1,0 @@
-# This function will prevent in-source builds
-function(AssureOutOfSourceBuilds)
-  # make sure the user doesn't play dirty with symlinks
-  get_filename_component(srcdir "${CMAKE_SOURCE_DIR}" REALPATH)
-  get_filename_component(bindir "${CMAKE_BINARY_DIR}" REALPATH)
-
-  # disallow in-source builds
-  if("${srcdir}" STREQUAL "${bindir}")
-		message(FATAL_ERROR "IN SOURCE BUILDS ARE NOT ALLOWED!\nDELETE CMakeCache.txt and CMakeFiles/ FROM THE ROOT DIRECTORY!\n")
-	endif()
-endfunction()
-
-AssureOutOfSourceBuilds()
